% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aez18.R
\docType{data}
\name{aez18}
\alias{aez18}
\title{SF with 18 Agroecological Zones}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 18 rows and 3 columns.
}
\source{
[\link{Micah}]
}
\usage{
aez18
}
\description{
The 18 Agroecoological Zones are based on GAEZ version XX [\link[=Ask
Micah for details and whether a vignette will be included how to
create the AEZ 18 by Country raster]{Ask Micah for details and whether a vignette will be included how to create the AEZ 18 by Country raster}].
}
\keyword{datasets}
