% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggr.R
\name{aggr}
\alias{aggr}
\title{Aggregate data to GTAP commodities}
\usage{
aggr(tmp_dir = getwd(), reg_country_conc_file, cpc_gsc3_concordance_file)
}
\arguments{
\item{tmp_dir}{Location of the tmp dir created using gtap_setup function. The default is the current working directory set by getwd()}

\item{reg_country_conc_file}{Location of a file linking the GADM countries to
the GTAP regions specified}

\item{cpc_gsc3_concordance_file}{File name of file containing concordance
which links CPC codes to their GSC3 GTAP sector codes}
}
\value{
Saves .rds files containing aggregated data.
}
\description{
This function aggregates the country-subnational division data to the
regions defined in the REG set made by the define_sets function. It also
aggregates crop production and the value of crop production to the 9
GTAP crop categories.
}
