% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{aggr_gtaplulc_3outputs}
\alias{aggr_gtaplulc_3outputs}
\title{Wrapper function to aggregate data by GTAP region and crop commodity
categories, create LULC .har files, create the input files, and then run
gtapland-ragg routines.}
\usage{
aggr_gtaplulc_3outputs(parent_dir, output_folder, direc_name)
}
\arguments{
\item{parent_dir}{Path to the parent directory where raw data, intermediate
outputs, and final datasets are stored}

\item{output_folder}{Desired name of folder containing 4 GTAP-LULC files:
BIOMEmod.txt (aggregation file), gtaplulcaez.har (land use and
land cover dataset), gtaplulcset.har (file containing the sets
used by GTAP-AEZ model). This folder is located in the "output_data"
folder within the specified parent directory.}

\item{direc_name}{Desired name of the final, aggregated database containing:
basedata.har, baserate.har, baseview.har, default.prm, inputs.zip, and
sets.har. This is output as a folder in the parent directory.}
}
\value{
Creates two folders. One contains the final, aggregated database and
a zipped folder with its input files. The second contains the GTAP-LULC
database for all GTAP regions (REG) defined by the definesets_lcover_luse
function.
}
\description{
Wrapper function to aggregate data by GTAP region and crop commodity
categories, create LULC .har files, create the input files, and then run
gtapland-ragg routines.
}
\examples{
aggr_gtaplulc_3outputs("myfiles/gtapfiles/", "GTAP_LULC_AEZ18_2017", "GTAP_AEZ18_2017")
}
