% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_gridded_df_to_subnatbound.R
\name{aggregate_gridded_df_to_subnatbound}
\alias{aggregate_gridded_df_to_subnatbound}
\title{Aggregate gridded land use and land cover to countries and
subnational boundaries of interest}
\usage{
aggregate_gridded_df_to_subnatbound(
  GADM_subnatbound_df,
  path.file.to.aggregate
)
}
\arguments{
\item{GADM_subnatbound_df}{Dataframe mapping the centroids of
(five-minute?) grid-cell coordinate to countries and geographic
boundaries of interest.}

\item{path.file.to.aggregate}{Character string with the path to a
dataframe with gridded information on land use (crop output and
harvested area) or land cover. These dataframes need to have
two columns labeled x an y with the coordinates of the
centroids of each five minutes gridcell on earth, under the
same projection GADM_subnatbound_df}
}
\value{
A data frame with the columns of \code{file.to.aggregate} (other
than x,y) aggregated to the countries and geographic boundaries
defined by GADM_subnatbound_df.
}
\description{
Aggregate gridded land use and land cover to countries and
subnational boundaries of interest
}
