% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_global_raster_properties.R
\name{apply_global_raster_properties}
\alias{apply_global_raster_properties}
\title{Apply Global Raster Properties}
\usage{
apply_global_raster_properties(input.raster, global.raster)
}
\arguments{
\item{input.raster}{a SpatRaster object}

\item{global.raster}{a SpatRaster object with a set of properties
that are imposed on each raster to ensure alignment once
rasters are converted to dataframes.}
}
\value{
a SpatRaster object with the following properties: extent
xmn = -180, xmx = 180, ymn = -90, ymx = 90: resolution =
c(0.08333333, 0.08333333), coordinate reference system: crs =
"+proj=longlat +datum=WGS84 +no_defs"
}
\description{
This function applies the properties of a global raster template to an input raster.
}
\examples{
\dontrun{
anyraster <- raster(matrix(runif(100), 10, 10))
goodraster <- apply_raster_properties(anyraster)
}
}
\seealso{
\link{global.raster} for details on the global raster template used.

\link{round_up_coordinates} to ensure that rasters can be merged with other rasters after converting them to dataframes.
}
