% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/apportion.iso.cropdata.to.subnatbound.and.gsc3.R
\name{apportion.iso.cropdata.to.subnatbound.and.gsc3}
\alias{apportion.iso.cropdata.to.subnatbound.and.gsc3}
\title{Make the Raster with Country-Subnational Boundaries}
\usage{
apportion.iso.cropdata.to.subnatbound.and.gsc3(
  GADM_subnatbound_df,
  year = 2017
)
}
\arguments{
\item{GADM_subnatbound_df}{A data frame with country (iso3 codes)
and subnational bounds created by
\code{\link[=make_subnatbound_gridded_dataframe]{make_subnatbound_gridded_dataframe()}}.}

\item{year}{Reference year of the database (defaults to 2017).}
}
\value{
A list of threee dataframes with crop quantity, value and
harvested area by GTAP region, subnational bound (e.g., 18
AEZs) and GTAP sector (gsc3).
}
\description{
Make the Raster with Country-Subnational Boundaries
}
\details{
....
}
\seealso{
\code{\link[=make_subnatbound_gridded_dataframe]{make_subnatbound_gridded_dataframe()}}
}
