% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_har_sets.R
\name{build_har_sets}
\alias{build_har_sets}
\title{Creates}
\usage{
build_har_sets(workdir_dir = getwd())
}
\arguments{
\item{workdir_dir}{Location of the workdir dir created using gtap_setup function. The default is the current working directory set by getwd()}
}
\value{
Outputs ....
}
\description{
Creates
}
