% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_land_cover_hectares.R
\name{calc_land_cover_hectares}
\alias{calc_land_cover_hectares}
\title{Calculate hectares in each land cover}
\usage{
calc_land_cover_hectares(workdir_dir = getwd())
}
\arguments{
\item{workdir_dir}{Location of the workdir dir created using gtap_setup function. The default is the current working directory set by getwd()}
}
\value{
Creates 7 rasters, one for each land cover type, containing hectares
in each grid cell
}
\description{
Calculates hectares using the rasters for each of the 7 land cover types
created by the combine_land_covers function
}
