% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_functions.R
\name{cleanup_intermediate_files}
\alias{cleanup_intermediate_files}
\title{Function to remove intermediate files}
\usage{
cleanup_intermediate_files(parent_dir)
}
\arguments{
\item{parent_dir}{Path to the parent directory}
}
\value{
Deletes files in the processed_data folder - NOTE - currently does
not delete any files.
}
\description{
Function to remove intermediate files
}
\examples{
cleanup_intermediate_files("mydir/gtapfiles/")
}
