% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_land_covers.R
\name{combine_land_covers}
\alias{combine_land_covers}
\title{Combine data to create 7 different land covers}
\usage{
combine_land_covers(
  cropland_rast_file,
  pasture_rast_file,
  workdir_dir = getwd()
)
}
\arguments{
\item{cropland_rast_file}{Path to the raster containing cropland area}

\item{pasture_rast_file}{Path to the raster containing pasture area}

\item{workdir_dir}{Location of the workdir dir created using gtap_setup function. The default is the current working directory set by getwd()}
}
\value{
Creates 7 rasters, one for each land cover type. Each raster
is a categorical raster indicating whether a cell is the respective
land cover
}
\description{
Uses cropland, pastureland, urbanland, and potential vegetation data to
create 7 land cover rasters. Cropland and pastureland are from Ramankutty
et al (2008). Urban land data are from Schneider, A., Friedl, M. A., &
Potere, D. (2010). Potential vegetation data are from Ramankutty and Foley
}
