% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shares.R
\name{create_shares}
\alias{create_shares}
\title{Create shares to disaggregate GTAP land rents}
\usage{
create_shares(monfreda_fao_concord_file, tmp_dir = getwd())
}
\arguments{
\item{monfreda_fao_concord_file}{File path to concordance linking monfreda
crop names to fao crop codes}

\item{tmp_dir}{Location of the tmp dir created using gtap_setup function. The default is the current working directory set by getwd()}
}
\value{
Saves .rds files containing share variables for country and
subnational area combinations.
}
\description{
This function creates country-BIOME share variables so the FAO production
data downloaded using the download_fao function can be disaggregated.
}
