% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_sets.R
\name{define_sets}
\alias{define_sets}
\title{Define sets}
\usage{
define_sets(reg_set_file, subnat_bound_file, workdir_dir = getwd())
}
\arguments{
\item{reg_set_file}{Data frame containing the set REG of GTAP regions.}

\item{subnat_bound_file}{Simple feature object containing the subnational boundaries (NV: Add a link to the function that creates this----would it be a good idea to have few of this included as defaults?)}

\item{workdir_dir}{Location of the workdir dir create using gtap_setup function. The default is the current working directory set by getwd()}
}
\value{
Sets for the final set of endowments (BIO_ENDW), 9 Crop commodity,
aggregates (CRP9) and 7 land cover types (LCOV) (NV: Add more details of what the ouput is and where the output is at. E.g., five text files with the sets ... and two Rds files with the sets ... in workdir/sets/ Also add links to the functions that use these sets to facilitate following the logic of the building process)
}
\description{
This function defines the sets used to generate agricultural
production data and split out land rents at a subnational level.There
are three sets used by the GTAP model which will be defined by the choices
made in this script:
\enumerate{
\item The region (REG) set elements will be defined.
\item A set of subnational boundaries (BIOMES) that sub-divide regions need
to be named and defined. For example, these biophysical boundaries may
be the 14 terrestrial BIOMES defined by Olson & Dinerstein (1998) for
the W.W.F. / Global 200 or they could be the 18 Agro-ecological zones
used in the GTAP-AEZ database.
\item The set of endowment commodities which includes countries' land endowments
by biophysical area, skilled labor, unskilled labor, capital, and
natural resources. The naming convention for this set is to use the
3 character name assigned to the set of biophysical boundaries followed
by "_ENDW".
}
}
\details{
The function has two arguments:
1. The file name of a .rds file containing the set of regions (with names
in column REG) and their numbers (in column REG_NUM).
2. The file name of an .rds file containing a simple features (sf) object
defining the sub-national boundaries and their names. The simple features
object containing the subnational boundaries needs to have a column named
"subnat_name" containing the names of the subnational areas,
another named "subnat_num" containing unique numbers for each,
and a "geometry" column that can be read by the 'sf' package.
}
