% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{definesets_lcover_luse}
\alias{definesets_lcover_luse}
\title{Wrapper function to run first three steps (Define sets, land cover, land use)}
\usage{
definesets_lcover_luse(
  parent_dir,
  reg_df,
  boundary_sf,
  hectare_rasters_needed,
  year_cropland_pasture_data
)
}
\arguments{
\item{parent_dir}{Path to the parent directory where raw data, intermediate
outputs, and final datasets stored}

\item{reg_df}{Data frame containing the set of regions}

\item{boundary_sf}{Simple feature object containing the subnational boundaries}

\item{hectare_rasters_needed}{Choice indicating whether the 7 land cover
rasters need to be created or whether they already exist ("YES" or "NO")}

\item{year_cropland_pasture_data}{Choice of whether to use 2000 or 2015 data
for cropland and pastureland area ("2000" or "2015")}
}
\value{
Sets for the final set of endowments (BIO_ENDW), 9 Crop commodity,
aggregates (CRP9) and 7 land cover types (LCOV), land cover data in
hectares for each combination of GADM country and subnational division,
files containing land use data for each combination of GADM
country and subnational division. Also produces raster indicating
concatenation of GADM countries and user-specified subnational divisions.
}
\description{
Wrapper function to run first three steps (Define sets, land cover, land use)
}
\examples{
definesets_lcover_luse("myfiles/gtapfiles/", GTAP_REG_dataframe, AEZ_shapefile, "YES", "2015")
}
