% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fao.R
\name{download_fao}
\alias{download_fao}
\title{Download FAO data}
\usage{
download_fao(gtap_year, tmp_dir = getwd())
}
\arguments{
\item{gtap_year}{Year of FAO data to be downloaded}

\item{tmp_dir}{Location of the tmp dir created using gtap_setup function. The default is the current working directory set by getwd()}
}
\value{
Saves .rds files containing production and price data
}
\description{
This function downloads FAO crop and livestock production data for a
specified year, as well as price data.
}
