% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_with_progress.R
\name{download_with_progress}
\alias{download_with_progress}
\title{Downloads a file and shows a progress bar}
\usage{
download_with_progress(url, destfile)
}
\arguments{
\item{url}{url from where to download the zipped data}

\item{destfile}{file to which the zipped data is written on disk}
}
\description{
Downloads a file and shows a progress bar
}
\seealso{
\link{getrawdata} for details on actions taken when the package is loaded.
}
