% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_bio_sets.R
\name{extract_bio_sets}
\alias{extract_bio_sets}
\title{Extract Sets with Geographies of Interest from the sf Boundaries}
\usage{
extract_bio_sets(subnat_bound_file)
}
\arguments{
\item{subnat_bound_file}{Simple feature object containing the subnational boundaries}
}
\value{
A character string with the names of the geographies of interest
}
\description{
Extract Sets with Geographies of Interest from the sf Boundaries
}
