% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{faodata_shares_sharedata}
\alias{faodata_shares_sharedata}
\title{Wrapper function to run steps 4-6 (download FAO, create shares, share out data)}
\usage{
faodata_shares_sharedata(parent_dir, gtap_year_choice)
}
\arguments{
\item{parent_dir}{Path to the parent directory where raw data, intermediate
outputs, and final datasets stored}

\item{gtap_year_choice}{Year of FAO data to be downloaded}
}
\value{
FAO crop and livestock production data and prices for a specified year,
and GADM country-BIOME share variables so the FAO production can be
disaggregated. The GADM country-BIOME share variables are then used to
disaggregate the national-level FAO data for the chosen GTAP reference year
by the user-specified subnational divisions.
}
\description{
Wrapper function to run steps 4-6 (download FAO, create shares, share out data)
}
\examples{
faodata_shares_sharedata("myfiles/gtapfiles/", "2017")
}
