% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getrawdata.r
\name{getrawdata}
\alias{getrawdata}
\title{Downloads GIS layers and FAO data}
\usage{
getrawdata(
  url = "https://www.gtap.agecon.purdue.edu/uploads/temp/gtapshape-raw-data.zip"
)
}
\arguments{
\item{url}{url from where to download the zipped data.}
}
\value{
A ~28GB folder in the current working directory named 'raw_data'.
}
\description{
The function downloads a large zipped data (~8GB) with
all the rasters for land cover, land use and ecological and
political boundaries and the FAO data that are needed to split
national land use, land cover and land rents. These data
(including sources) are documented in the companion
article. The data is stored in the GTAP servers.

This function is ran automatically when the package is installed
for the first time and can be ran by itself if needed. The
function uses \link{download_with_progress} to handle the
download of the file and showing a progress bar.
}
\examples{
\dontrun{getrawdata(url="https://www.gtap.agecon.purdue.edu/uploads/temp/FlexSpatialAgg.zip")}

}
\seealso{
\link{.onAttach} for details on actions taken when the package is loaded.
}
