% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTAP.ruminant.prices.R
\docType{data}
\name{GTAP.ruminant.prices}
\alias{GTAP.ruminant.prices}
\title{GTAP Ruminant Prices}
\format{
A data frame with the following columns:
\describe{
\item{iso3}{Country}
\item{use}{Ruminant species}
\item{unit}{USD/head}
\item{value}{Value}
}
}
\source{
"The script to generate this dataset can be found in the
\code{data-raw} folder on GitHub:
\url{https://github.com/nvilloria/gtapshape/tree/main/data-raw}"
}
\usage{
GTAP.ruminant.prices
}
\description{
This dataset contains GTAP ruminant prices. These livestock
prices are taken from the Land Use and Land Cover (LULC) Data associated with
the version 11c GTAP-AEZ database. The construction of the version 11
GTAP-AEZ LULC database is described in \href{https://www.gtap.agecon.purdue.edu/resources/res_display.asp?RecordID=7407}{Baldoz and Corong (2025)}.
To recover prices, we sum up the value of livestock production (in 1,000 USD)
and quantity of livestock production (in 1,000 head) from the v11 GTAP-AEZ
LULC database by GTAP region. Then we divide the region-level value of
production by the quantity of production to recover the price.
}
\keyword{datasets}
