% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtaplulc.R
\name{gtaplulc}
\alias{gtaplulc}
\title{Create gtaplulcaez.har file}
\usage{
gtaplulc(
  workdir_dir = getwd(),
  gempack_lulc_dir,
  gempack_timber_dir,
  gempack_rentcalc_dir
)
}
\arguments{
\item{workdir_dir}{Location of the workdir dir created using gtap_setup function. The default is the current working directory set by getwd()}

\item{gempack_lulc_dir}{Location of the directory where inputs and outputs
to LULC GEMPACK executables are stored.}

\item{gempack_timber_dir}{Location of the directory where inputs and outputs
for the GEMPACK timberland rents executables are}

\item{gempack_rentcalc_dir}{Location of the directory where inputs and outputs
for the GEMPACK rentcalc executables are}
}
\value{
Outputs gtaplulcaez.har file.
}
\description{
This function uses the user defined sets to create a .har file using MODHAR,
then it reads in the gempack text files created by gpack_txt
}
