% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file.R
\name{input_file}
\alias{input_file}
\title{Create aggregation input file}
\usage{
input_file(
  tmp_dir = getwd(),
  gempack_rentcalc_dir,
  input_aggr_file,
  input_aggr_file_dir
)
}
\arguments{
\item{tmp_dir}{Location of the tmp dir created using gtap_setup function. The default is the current working directory set by getwd()}

\item{gempack_rentcalc_dir}{Location of the directory where inputs and outputs}

\item{input_aggr_file}{Aggregation file (.txt) with mapping for regions and
commodity groups.}

\item{input_aggr_file_dir}{Location of the directory containing the input file
used to aggregate LULC databse to desired regions, commodity groups, etc.}
}
\value{
Outputs BIOMEmod.txt aggregation file.
}
\description{
This function creates an input file for aggregating the data. It modifies the
end of the BIOMEmod_base.txt file to add in the sub-national boundary sets.
}
