% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faostat.crop.production.by.iso.R,
%   R/iso.crop.production.2011.2022.R
\docType{data}
\name{iso.crop.production.2011.2022}
\alias{iso.crop.production.2011.2022}
\title{My Example Dataset}
\format{
A data frame with 5 rows and 2 variables:
\describe{
\item{id}{Integer ID}
\item{name}{Character string of the person's name}
}

A data frame with 5 rows and 2 variables:
\describe{
\item{id}{Integer ID}
\item{name}{Character string of the person's name}
}
}
\source{
Created for demonstration purposes

Created for demonstration purposes
}
\usage{
iso.crop.production.2011.2022

iso.crop.production.2011.2022
}
\description{
A dataset containing example data with IDs and names.

A dataset containing example data with IDs and names.
}
\keyword{datasets}
