% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.land.cover.2011.2022.R
\docType{data}
\name{iso.land.cover.2011.2022}
\alias{iso.land.cover.2011.2022}
\title{Annual Land Covers by Country (2011-2022)}
\format{
A data frame with 5 rows and 2 variables:
\describe{
\item{iso3}{3-digit ISO country code.}
\item{year}{Data is distributed for 2011-2022.}
\item{gsc3}{cropland and pasture, both reported annually by FAOSTAT; urban is fixed over time;  forest, grassland, shrubland, and other are estimates obtained by sharing out unclassified land in FAOSTAT using gridded land cover data for 2000.
}
}
\source{
[\link{Micah: Add source for gridded land cover data, including urban}]. The step-by-step procedures to create this dataset can be found in the \link[=Country_level_landcover_ts]{vignette about process annual, country-level land cover data}.
}
\usage{
iso.land.cover.2011.2022
}
\description{
A dataframe containing annual cropland and pastures reported by FAOSTAT, and estimated annual forests, grasslands, shrublands, and other obtained by sharing out unclassified lands in faostat using the data on land cover for 2000 [\link{Micah: Add source for land cover data}].
}
\keyword{datasets}
