% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.cover.concordance.R
\docType{data}
\name{land.cover.concordance}
\alias{land.cover.concordance}
\title{Land Cover Concordance}
\format{
A data frame with 7 rows and 2 columns:
\describe{
\item{gsc3}{Global Land Cover Share category.}
\item{use}{Corresponding 'use' category within the package.}
}
}
\source{
The script to generate this dataset can be found in the
\code{data-raw} folder on GitHub:
\url{https://github.com/nvilloria/gtapshape/tree/main/data-raw}
}
\usage{
land.cover.concordance
}
\description{
This dataset provides a concordance between gsc3 (Global Land Cover Share) categories and the 'use' categories used in this package.
}
\keyword{datasets}
