% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_cover.R
\name{land_cover}
\alias{land_cover}
\title{Land cover}
\usage{
land_cover(gadm_rast_file, subnat_bound_file, workdir_dir = getwd())
}
\arguments{
\item{gadm_rast_file}{File name for raster containing GADM countries}

\item{subnat_bound_file}{Simple feature object containing the subnational boundaries}

\item{workdir_dir}{Location of the workdir dir created using gtap_setup function. The default is the current working directory set by getwd()}
}
\value{
Land cover data in hectares for each combination of GADM country and
subnational division
}
\description{
This function creates rasters for potential vegetation and urban land cover,
and then combines them with cropland and pastureland rasters. Next, it
creates 7 rasters showing the hectares in each land cover. Next, output
the number of hectares in each land cover by country-subnational boundary
combination.
}
