% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_country_biome_raster.R
\name{make_country_biome_raster}
\alias{make_country_biome_raster}
\title{Make the Raster with Country-Geographies Boundaries}
\usage{
make_country_biome_raster(subnat_bound_file)
}
\arguments{
\item{subnat_bound_file}{Simple feature object containing the subnational boundaries}
}
\value{
A SpatRaster object with boundaries for countries and the geographies of interest.
}
\description{
Make the Raster with Country-Geographies Boundaries
}
