% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_country_biome_gridded_dataframe.R
\name{make_subnatbound_gridded_dataframe}
\alias{make_subnatbound_gridded_dataframe}
\title{Creates a Dataframe with Coordinates for Country-Geographies Boundaries}
\usage{
make_subnatbound_gridded_dataframe(subnat_bound_file)
}
\arguments{
\item{subnat_bound_file}{Simple feature object containing the
subnational boundaries}
}
\value{
A dataframe with the xy coordinates of each earth gridcell
and a column with the categoris for the different
countrry-geopgraphies defined in subnat_bound_file.
}
\description{
The resulting dataframe is used to aggregated the gridded datasets
on land use and land cover to the country-geographies of
interest. The function is a wrapper of
\link{make_country_biome_raster} and
\link{round_up_coordinates}.
}
