% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_subnatbound_raster.R
\name{make_subnatbound_raster}
\alias{make_subnatbound_raster}
\title{Make the Raster with Country-Subnational Boundaries}
\usage{
make_subnatbound_raster(subnat_bound_file = "aez18")
}
\arguments{
\item{subnat_bound_file}{Simple feature object containing the
subnational boundaries. The default is \link{aez18}. \link{biomes14} is
available for lazy load,
e.g. subnat_bound_file='biomes14'. Other sf objects with
subnational boundaries can be passed on by using the full path
of the file.}
}
\value{
A SpatRaster object with boundaries for countries and the
geographies of interest.
}
\description{
Make the Raster with Country-Subnational Boundaries
}
\seealso{
\code{\link[=make_subnatbound_gridded_dataframe]{make_subnatbound_gridded_dataframe()}} which invokes this
function before creating a dataframe with XY coordinates and
the combinations of countries and subnational boundaries.
}
