% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_workdir.R
\name{make_workdir}
\alias{make_workdir}
\title{Setup gtapshape}
\usage{
make_workdir(workdir_dir = getwd())
}
\arguments{
\item{tmp_dir}{Location of the tmp dir created using gtap_setup function. The default is the current working directory set by getwd()}
}
\value{
Creates workdir directory along with its sub-directories
}
\description{
This function creates directories in the user-specified locations.
}
