% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_to_zero.R
\name{na_to_zero}
\alias{na_to_zero}
\title{NA to zero}
\usage{
na_to_zero(a)
}
\arguments{
\item{a}{A terra SpatRaster}
}
\value{
A terra SpatRaster where cells with NA values now have zeroes
}
\description{
This function replaces NA values in a terra SpatRaster with zeroes
}
