% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg_to_zero.R
\name{neg_to_zero}
\alias{neg_to_zero}
\title{Negative to zero}
\usage{
neg_to_zero(a)
}
\arguments{
\item{a}{A terra SpatRaster object}
}
\value{
Returns the terra SpatRaster with negative values replaced by zeroes
}
\description{
Replace negative raster values with zeros.
}
