% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.gsc3.by.iso.R
\name{order.gsc3.by.iso}
\alias{order.gsc3.by.iso}
\title{Orders gsc3.by.iso output to be consistent with set order from GTAP
base data}
\usage{
order.gsc3.by.iso(
  gsc3.by.iso,
  gtap_basedatasets_file,
  subnat_bound_file = "aez18"
)
}
\arguments{
\item{gsc3.by.iso}{List containing shared out land use and land cover data}

\item{gtap_basedatasets_file}{Base GTAP database set file}

\item{subnat_bound_file}{SF file with subnational
boundaries. Defaults to the 18 Agroecological Zones used in version 11
of the GTAP-AEZ LULC database by \href{https://www.gtap.agecon.purdue.edu/resources/res_display.asp?RecordID=7407}{Baldoz and Corong (2025)}.
The \code{vignette("create.18.aez.shapefile", package = "gtapshape")} contains
the code which creates the 18 AEZ shapefile.}
}
\description{
Orders gsc3.by.iso output to be consistent with set order from GTAP
base data
}
