% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential_vegetation.R
\name{potential_vegetation}
\alias{potential_vegetation}
\title{Potential vegetation}
\usage{
potential_vegetation(pot_veg_file, tmp_dir = getwd())
}
\arguments{
\item{pot_veg_file}{Path to the raster file indicating potential vegetation classes}

\item{tmp_dir}{Location of the tmp dir create using gtap_setup function. The default is the current working directory set by getwd()}
}
\value{
Saves 4 rasters for the "Forests", "Shrubland",
"Savanna + Grassland", and "Other Lands" potential vegetation types from
the original data produced by Ramankutty & Foley.
}
\description{
This function creates rasters for potential vegetation.
}
