% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.concordance.R
\docType{data}
\name{regional.concordance}
\alias{regional.concordance}
\title{My Example Dataset}
\format{
A data frame with 5 rows and 2 variables:
\describe{
\item{id}{Integer ID}
\item{name}{Character string of the person's name}
}
}
\source{
Created for demonstration purposes
}
\usage{
regional.concordance
}
\description{
A dataset containing example data with IDs and names.
}
\keyword{datasets}
