% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_up_coordinates.R
\name{round_up_coordinates}
\alias{round_up_coordinates}
\title{Round Up Coordinates of Gridded Dataframes}
\usage{
round_up_coordinates(raster.df)
}
\arguments{
\item{raster.df}{a dataframe with x,y coordinates}
}
\value{
a dataframe with coordinates rounded up to six digits
}
\description{
Round Up Coordinates of Gridded Dataframes
}
\examples{
\dontrun{
anyraster <- raster(matrix(runif(100), 10, 10))
goodraster <- apply_raster_properties(anyraster)
}
}
\seealso{
\link{apply_global_raster_properties} which usually preceds the
application of this function.
}
