% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_3final_outputs.R
\name{save_3final_outputs}
\alias{save_3final_outputs}
\title{Save LULC final outputs with aggregation .txt file}
\usage{
save_3final_outputs(gempack_rentcalc_dir, out_folder_name, out_folder_dir)
}
\arguments{
\item{gempack_rentcalc_dir}{Location of the directory where inputs and outputs}

\item{out_folder_name}{Desired name of folder containing final outputs}

\item{out_folder_dir}{Location where the output folder should be saved}
}
\value{
Outputs a folder containing 4 files: BIOMEmod.txt (aggregation file),
gtaplulcaez.har (land use and land cover dataset), gtaplulcset.har (file
containing the sets used by GTAP-AEZ model).
}
\description{
Define function to create new folder and move output files there.
Function takes 2 arguments: the parent directory containing the
GTAP_BIOMES folder, and the desired name of the folder which will
contain the 3 output files.
For example:
parent_dir <- "C:/Users/Micah/Dropbox/CGE_BIOMES_Carbon_2024/"
out_folder_name <- "GTAP_BRAZD7_beefandsoy75"
}
