% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_gtapshape.R
\name{setup_gtapshape}
\alias{setup_gtapshape}
\title{Setup gtapshape}
\usage{
setup_gtapshape(workdir_dir = getwd())
}
\arguments{
\item{tmp_dir}{Location of the tmp dir created using gtap_setup function. The default is the current working directory set by getwd()}
}
\value{
Creates workdir directory along with its sub-directories
}
\description{
This function creates directories in the user-specified locations.
}
