% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share.out.sectors.to.subnatbound.R
\name{share.out.sectors.to.subnatbound}
\alias{share.out.sectors.to.subnatbound}
\title{Dissagregates national data by subantional boundary and aggregates to GTAP regions}
\usage{
share.out.sectors.to.subnatbound(iso.data, shares, units, sector)
}
\arguments{
\item{iso.data}{A dataframe with national data for a given
year. The structure (column names) of these dataframes must be
identical to the datasets iso.crop.production.2011.2022,
iso.crop.outputvalue.2011.2022, iso.crop.harea.2011.2022,
iso.gsc3lstk.heads.2011.2022,
iso.gsc3lstk.outputvalue.2011.2022 and
iso.land.cover.2011.2022, all distributed with \code{gtrapshape} and
available for lazy loading.}

\item{shares}{A dataframe created by \code{subnatbound.shares()},}

\item{units}{The units of the physical data. See vignette
\code{build.gtapdabase.for.year.and.from.sf} or
\code{build.dbase.from.sf()} for a programmatic way to get this from
the iso.data.}

\item{sector}{For crops is 'item_code_cpc'; 'gsc3' for the
rest. See vignette \code{build.gtapdabase.for.year.and.from.sf} or
\code{build.dbase.from.sf()} for a programmatic way to get this from
the iso.data.}
}
\description{
Dissagregates national data by subantional boundary and aggregates to GTAP regions
}
