% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_data.R
\name{share_data}
\alias{share_data}
\title{Share out land rents}
\usage{
share_data(
  lvstk_prices_4spec_file,
  cpc_gsc3_concordance_file,
  tmp_dir = getwd()
)
}
\arguments{
\item{lvstk_prices_4spec_file}{File name of file containing prices for four
livestock species. Should have five columns: "QLIVE_18", "X1.BUFFALOES",
"X2.CATTLE", "X3.GOATS", "X4.SHEEP")}

\item{cpc_gsc3_concordance_file}{File name of file containing concordance
which links CPC codes to their GSC3 GTAP sector codes}

\item{tmp_dir}{Location of the tmp dir created using gtap_setup function. The default is the current working directory set by getwd()}
}
\value{
Saves .rds files containing land cover, crop production, livestock
production, and value of production data for country and subnational
area combinations.
}
\description{
This function uses the country-BIOME share variables to disaggregate the fao
data for the chosen GTAP reference year. National level data from the FAO
are first split by subnational division, and then we aggregate to the REG
-subnational division level using the aggr function.
}
