% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardGTAPsets.R
\name{standardGTAPsets}
\alias{standardGTAPsets}
\title{Title}
\usage{
standardGTAPsets(standardsets = NULL)
}
\arguments{
\item{standardsets}{Regional and sectoral sets used by the standard GTAP data. The default is V10. \link{NV: Micah is this right?}}
}
\description{
Title
}
