% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnatbound.shares.R
\name{subnatbound.shares}
\alias{subnatbound.shares}
\title{Calculate the subnational boundary shares of land uses and covers `}
\usage{
subnatbound.shares(subnatbound.use, sector)
}
\arguments{
\item{subnatbound.use}{Dataframe with the value of different uses
by country (iso3), subnational bound, and use (either crops,
livestock and land cover). This dataframe must have a target
sector to which the uses are aggregated.}

\item{sector}{The sector to which the uses are aggregated: CPC for
the case of crops and GSC3 for livestock and land covers.}
}
\description{
Calculate the subnational boundary shares of land uses and covers `
}
\seealso{
\code{\link[=all_uses_by_subnatbound]{all_uses_by_subnatbound()}} creates dataframes from
gridded data aggregate to subnational boundes. These dataframes
need to be merged with a sectoral concordance, eithr
\code{crop.concordance}, \code{gridded.livestock.concordance} or
land.cover.concordance in order to have a target sectors
}
