% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urban_land_cover.R
\name{urban_land_cover}
\alias{urban_land_cover}
\title{Urban land cover}
\usage{
urban_land_cover(bip_file, hdr_file, cropland_file, tmp_dir = getwd())
}
\arguments{
\item{bip_file}{Path to the bip file containing urban land cover data}

\item{hdr_file}{Path to the hdr file for the urban land cover data}

\item{cropland_file}{Path to the file containing cropland data}

\item{tmp_dir}{Location of the tmp dir created using gtap_setup function. The default is the current working directory set by getwd()}
}
\value{
Creates urban cover raster
}
\description{
Reading in the urban cover raster - Schneider, A., Friedl, M. A., & Potere,
D. (2010). Mapping global urban areas using MODIS 500-m data
}
