% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.gtaplulc.har.R
\name{write.gtaplulc.har}
\alias{write.gtaplulc.har}
\title{Write HAR file with physical land use and cover for 160 GTAP regions and subantional bounds}
\usage{
write.gtaplulc.har(gsc3.by.iso, file = "gtaplulc.har")
}
\arguments{
\item{gsc3.by.iso}{A list with the dataframes of physical land use
and cover data by GTAP region and subnational boundaries. Each
dataframe must have columns labeled 'reg', 'subnatbound\verb{, }gsc3\verb{, and value.  This function is a wrapper of the steps needed to write har files using }HARr::write_har()`.}

\item{file}{Name of the har file with the aggregated physical
data. Should have an extension. Defaults to "gtaplulc.har"}
}
\description{
Write HAR file with physical land use and cover for 160 GTAP regions and subantional bounds
}
