% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.gtaplulcsets.har.R
\name{write.gtaplulcsets.har}
\alias{write.gtaplulcsets.har}
\title{Write HAR file containing the sets needed to read in the land use and
land cover data. These need to be defined in order so they are
consistent with the GTAP base database.}
\usage{
write.gtaplulcsets.har(
  gtap_basedatasets_file,
  subnat_bound_file = "aez18",
  file = "gtaplulcsets.har"
)
}
\arguments{
\item{gtap_basedatasets_file}{Base GTAP database set file}

\item{subnat_bound_file}{SF file with subnational
boundaries. Defaults to the 18 Agroecological Zones used in version 11
of the GTAP-AEZ LULC database by \href{https://www.gtap.agecon.purdue.edu/resources/res_display.asp?RecordID=7407}{Baldoz and Corong (2025)}.
The \code{vignette("create.18.aez.shapefile", package = "gtapshape")} contains
the code which creates the 18 AEZ shapefile.}

\item{file}{Name of the output HAR file. Defaults to 'gtaplulcsets.har'.}
}
\description{
Write HAR file containing the sets needed to read in the land use and
land cover data. These need to be defined in order so they are
consistent with the GTAP base database.
}
