% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shares.R, R/create_shares2.R
\name{create_shares}
\alias{create_shares}
\title{Create shares to disaggregate GTAP land rents}
\usage{
create_shares(monfreda_fao_concord_file, workdir_dir = getwd())

create_shares(monfreda_fao_concord_file, workdir_dir = getwd())
}
\arguments{
\item{monfreda_fao_concord_file}{File path to concordance linking monfreda
crop names to fao crop codes}

\item{workdir_dir}{Location of the workdir dir created using gtap_setup function. The default is the current working directory set by getwd()}
}
\value{
Saves  the three .rds files below to ./workdir/. These files contain share variables for country and
subnational area combinations.

gadm_bio_crop_prod_shares.rds

gadm_bio_landcov_shares.rds

\link{NV: What are these data for and which function uses them?}

Saves  the three .rds files below to ./workdir/. These files contain share variables for country and
subnational area combinations.

gadm_bio_crop_prod_shares.rds

gadm_bio_landcov_shares.rds

\link{NV: What are these data for and which function uses them?}
}
\description{
This function creates country-BIOME share variables so the FAO production
data downloaded using the download_fao function can be disaggregated.

This function creates country-BIOME share variables so the FAO production
data downloaded using the download_fao function can be disaggregated.
}
