% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpack_txt.R
\name{gpack_txt}
\alias{gpack_txt}
\title{Output aggregated data as GEMPACK txt files}
\usage{
gpack_txt(tmp_dir = getwd(), cpc_gsc3_concordance_file)
}
\arguments{
\item{tmp_dir}{Location of the tmp dir created using gtap_setup function. The default is the current working directory set by getwd()}

\item{cpc_gsc3_concordance_file}{File name of file containing concordance
which links CPC codes to their GSC3 GTAP sector codes}
}
\value{
Saves .txt files in GEMPACK txt file format containing aggregated data.
}
\description{
This function outputs the aggregated data from the "aggr" command as
GEMPACK text files so they can be read into GTAP analyses.
}
