% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTAP.ruminant.prices.R
\docType{data}
\name{GTAP.ruminant.prices}
\alias{GTAP.ruminant.prices}
\title{GTAP Ruminant Prices}
\format{
A data frame with the following columns:
\describe{
\item{year}{Year of data}
\item{price}{Ruminant price}
}
}
\source{
The script to generate this dataset can be found in the
\code{data-raw} folder on GitHub:
\url{https://github.com/nvilloria/gtapshape/tree/main/data-raw}
}
\usage{
GTAP.ruminant.prices
}
\description{
This dataset contains GTAP ruminant prices.
}
\keyword{datasets}
