% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_subnatbound_gridded_dataframe.R
\name{make_subnatbound_gridded_dataframe}
\alias{make_subnatbound_gridded_dataframe}
\title{Creates a Dataframe with Coordinates for Country-Geographies Boundaries}
\usage{
make_subnatbound_gridded_dataframe(subnat_bound_file = "aez18")
}
\arguments{
\item{subnat_bound_file}{Simple feature object containing the
subnational boundaries. The default subnational boundaries are
Country by AEZ 18 (see \link{aez18}). \link{biomes14} is available for
lazy load, e.g. subnat_bound_file='biome14'. Other sf objects
with subnational boundaries can be passed on by using the full
path of the file.}
}
\value{
A dataframe with the xy coordinates of each earth gridcell
and a column with the categoris for the different
countrry-geopgraphies defined in subnat_bound_file.
}
\description{
The resulting dataframe is used to aggregated the gridded datasets
on land use and land cover to the country-geographies of
interest. The function is a wrapper of
\link{make_country_biome_raster} and
\link{round_up_coordinates}.
}
\seealso{
\code{\link[=make_subnatbound_raster]{make_subnatbound_raster()}} which produces the raster with
country-subnational boundaries by intersecting polygon files
with countries and subnational boundaries.
}
