% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.gsc3.by.iso.R
\name{order.gsc3.by.iso}
\alias{order.gsc3.by.iso}
\title{Orders gsc3.by.iso output to be consistent with set order from GTAP
base data}
\usage{
order.gsc3.by.iso(
  gsc3.by.iso,
  gtap_basedatasets_file,
  subnat_bound_file = "aez18"
)
}
\arguments{
\item{gsc3.by.iso}{List containing shared out land use and land cover data}

\item{gtap_basedatasets_file}{Base GTAP database set file}

\item{subnat_bound_file}{SF file with subnational
boundaries. Defaults to the 18 Agroecological Zones (Micah:
Version AND perhaps refer to vignette?)#'}
}
\description{
Orders gsc3.by.iso output to be consistent with set order from GTAP
base data
}
