% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_data.R
\name{share_data}
\alias{share_data}
\title{Share out land rents}
\usage{
share_data(
  lvstk_prices_4spec_file,
  cpc_gsc3_concordance_file,
  workdir_dir = getwd()
)
}
\arguments{
\item{lvstk_prices_4spec_file}{File name of file containing prices for four
livestock species. Should have five columns: "QLIVE_18", "X1.BUFFALOES",
"X2.CATTLE", "X3.GOATS", "X4.SHEEP")}

\item{cpc_gsc3_concordance_file}{File name of file containing concordance
which links CPC codes to their GSC3 GTAP sector codes}

\item{workdir_dir}{Location of the workdir dir created using gtap_setup function. The default is the current working directory set by getwd()}
}
\value{
Saves the nine .rds files below in
./workdir/gtap_ref_year. These rsd files containing land cover,
crop production, livestock production, and value of production
data for country and subnational area combinations.

gadm_bio_crop_harvarea.rds *

gadm_bio_crop_production.rds *

gadm_bio_livestock_valprod_ctlrmkwol.rds *

gadm_bio_livestock_valprod_4species.rds

gadm_bio_livestock_production_4species.rds

gadm_bio_livestock_numanimals.rds *

gadm_bio_land_cover.rds *

gadm_bio_vallvstkprod_rmk_wol.rds

gadm_bio_val_crop_production.rds *

The data with a star * are an input to \link{aggr}.
}
\description{
This function uses the country-BIOME share variables to disaggregate the fao
data for the chosen GTAP reference year. National level data from the FAO
are first split by subnational division, and then we aggregate to the REG
-subnational division level using the aggr function.
}
