% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.gtaplulcagg.txt.R
\name{write.gtaplulcagg.txt}
\alias{write.gtaplulcagg.txt}
\title{Create aggregation textfile based on subnatbound choice}
\usage{
write.gtaplulcagg.txt(
  subnat_bound_file = "aez18",
  base_aggr_file = system.file("agg_templates", "gtapv11c_aggr_base_file.txt", package =
    "gtapshape"),
  textfilename = "gtaplulcagg.txt"
)
}
\arguments{
\item{subnat_bound_file}{SF file with subnational
boundaries. Defaults to the 18 Agroecological Zones (Micah:
Version AND perhaps refer to vignette?)#'}

\item{base_aggr_file}{Text file containing how commodities will are mapped to
larger categories, how regions will be aggregated, etc. NOTE - This file
will be overwritten when this function runs.}

\item{file}{Name of the output aggregation file. Defaults to 'gtaplulcagg.txt'.}
}
\description{
Create aggregation textfile based on subnatbound choice
}
